+++ cmd/edgectl/aes_install.go
@@ -20,10 +20,17 @@ import (
 	"syscall"
 	"time"
 
+	"github.com/pkg/browser"
+
 	"github.com/datawire/ambassador/pkg/k8s"
 	"github.com/datawire/ambassador/pkg/supervisor"
+	"github.com/gookit/color"
 	"github.com/pkg/errors"
 	"github.com/spf13/cobra"
+	k8sTypesMetaV1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	k8sVersion "k8s.io/apimachinery/pkg/version"
+	k8sClientCoreV1 "k8s.io/client-go/kubernetes/typed/core/v1"
+	"k8s.io/client-go/rest"
 )
 
 func aesInstallCmd() *cobra.Command {
@@ -71,14 +78,23 @@ func getEmailAddress(defaultEmail string, log *log.Logger) string {
 			return text
 		}
 
-		fmt.Printf("Sorry, %q does not match our email address filter.\n", text)
+		fmt.Printf("Sorry, %q does not appear to be a valid email address.  Please check it and try again.\n", text)
 	}
 }
 
 func aesInstall(cmd *cobra.Command, args []string) error {
+	skipReport, _ := cmd.Flags().GetBool("no-report")
 	verbose, _ := cmd.Flags().GetBool("verbose")
 	kcontext, _ := cmd.Flags().GetString("context")
 	i := NewInstaller(verbose)
+
+	// If Scout is disabled (environment variable set to non-null), inform the user.
+	if i.scout.Disabled() {
+		i.show.Printf(phoneHomeDisabled)
+	}
+
+	// Both printed and logged when verbose (Installer.log is responsible for --verbose)
+	i.log.Printf(fmt.Sprintf(installAndTraceIDs, i.scout.installID, i.scout.metadata["trace_id"]))
 
 	sup := supervisor.WithContext(i.ctx)
 	sup.Logger = i.log
@@ -103,9 +119,16 @@ func aesInstall(cmd *cobra.Command, args []string) error {
 		Requires: []string{"signal"},
 		Work: func(p *supervisor.Process) error {
 			defer i.Quit()
-			return i.Perform(kcontext)
+			result := i.Perform(kcontext)
+			i.ShowResult(result)
+			return result.Err
 		},
 	})
+
+	// Don't allow messages emitted while opening the browser to mess up our
+	// carefully-crafted terminal output
+	browser.Stdout = ioutil.Discard
+	browser.Stderr = ioutil.Discard
 
 	runErrors := sup.Run()
 	if len(runErrors) > 1 { // This shouldn't happen...
@@ -114,6 +137,9 @@ func aesInstall(cmd *cobra.Command, args []string) error {
 		}
 	}
 	if len(runErrors) > 0 {
+		if !skipReport {
+			i.generateCrashReport(runErrors[0])
+		}
 		i.show.Println()
 		i.show.Printf("Full logs at %s\n\n", i.logName)
 		return runErrors[0]
@@ -183,7 +209,7 @@ func (i *Installer) loopUntil(what string, how func() error, lc *loopConfig) err
 		// Wait and try again
 		select {
 		case <-progTimer.C:
-			i.show.Printf("(waiting for %s)", what)
+			i.show.Printf("   Still waiting for %s. (This may take a minute.)", what)
 		case <-time.After(lc.sleepTime):
 			// Try again
 		case <-ctx.Done():
@@ -197,37 +223,80 @@ func (i *Installer) loopUntil(what string, how func() error, lc *loopConfig) err
 // the Pod is Running (though not necessarily Ready). This should be good enough
 // to report the "deploy" status to metrics.
 func (i *Installer) GrabAESInstallID() error {
-	// Note: This is brittle. If there is more than one pod, this returns all
-	// the pod names squashed together. This should not occur in normal usage,
-	// i.e. when installing a single version of AES, maybe repeatedly.
-	pod, err := i.CaptureKubectl("get AES pod", "", "-n", "ambassador", "get", "pods", "-l", "service=ambassador", "-o", "go-template={{range .items}}{{.metadata.name}}{{end}}")
+	aesImage := "quay.io/datawire/aes:" + i.version
+	podName := ""
+	containerName := ""
+	podInterface := i.coreClient.Pods("ambassador") // namespace
+	i.log.Print("> k -n ambassador get po")
+	pods, err := podInterface.List(k8sTypesMetaV1.ListOptions{})
 	if err != nil {
 		return err
 	}
-	clusterID, err := i.CaptureKubectl("get cluster ID", "", "-n", "ambassador", "exec", pod, "python3", "kubewatch.py")
+
+	// Find an AES Pod
+PodsLoop:
+	for _, pod := range pods.Items {
+		i.log.Print("  Pod: ", pod.Name)
+	ContainersLoop:
+		for _, container := range pod.Spec.Containers {
+			// Avoid matching the Traffic Manager (container.Command == ["traffic-proxy"])
+			i.log.Printf("       Container: %s (image: %q; command: %q)", container.Name, container.Image, container.Command)
+			if container.Image != aesImage || container.Command != nil {
+				continue
+			}
+			// Avoid matching the Traffic Agent by checking for
+			// AGENT_SERVICE in the environment. This is how Ambassador's
+			// Python code decides it is running as an Agent.
+			for _, envVar := range container.Env {
+				if envVar.Name == "AGENT_SERVICE" && envVar.Value != "" {
+					i.log.Printf("                  AGENT_SERVICE: %q", envVar.Value)
+					continue ContainersLoop
+				}
+			}
+			i.log.Print("       Success")
+			podName = pod.Name
+			containerName = container.Name
+			break PodsLoop
+		}
+	}
+	if podName == "" {
+		return errors.New("no AES pods found")
+	}
+
+	// Retrieve the cluster ID
+	clusterID, err := i.CaptureKubectl("get cluster ID", "", "-n", "ambassador", "exec", podName, "-c", containerName, "python3", "kubewatch.py")
 	if err != nil {
 		return err
 	}
+	i.clusterID = clusterID
 	i.SetMetadatum("Cluster ID", "aes_install_id", clusterID)
 	return nil
 }
 
-// GrabLoadBalancerIP return's the AES service load balancer's IP address
-func (i *Installer) GrabLoadBalancerIP() (string, error) {
-	ipAddress, err := i.CaptureKubectl("get IP address", "", "get", "-n", "ambassador", "service", "ambassador", "-o", `go-template={{range .status.loadBalancer.ingress}}{{print .ip "\n"}}{{end}}`)
+// GrabLoadBalancerAddress retrieves the AES service load balancer's address (IP
+// address or hostname)
+func (i *Installer) GrabLoadBalancerAddress() error {
+	serviceInterface := i.coreClient.Services("ambassador") // namespace
+	service, err := serviceInterface.Get("ambassador", k8sTypesMetaV1.GetOptions{})
 	if err != nil {
-		return "", err
+		return err
 	}
-	ipAddress = strings.TrimSpace(ipAddress)
-	if ipAddress == "" {
-		return "", errors.New("empty IP address")
+	for _, ingress := range service.Status.LoadBalancer.Ingress {
+		if net.ParseIP(ingress.IP) != nil {
+			i.address = ingress.IP
+			return nil
+		}
+		if ingress.Hostname != "" {
+			i.address = ingress.Hostname
+			return nil
+		}
 	}
-	return ipAddress, nil
+	return errors.New("no address found")
 }
 
 // CheckAESServesACME performs the same checks that the edgestack.me name
-// service performs against the AES load balancer IP
-func (i *Installer) CheckAESServesACME(ipAddress string) (err error) {
+// service performs against the AES load balancer host
+func (i *Installer) CheckAESServesACME() (err error) {
 	defer func() {
 		if err != nil {
 			i.log.Print(err.Error())
@@ -235,7 +304,7 @@ func (i *Installer) CheckAESServesACME(ipAddress string) (err error) {
 	}()
 
 	// Verify that we can connect to something
-	resp, err := http.Get("http://" + ipAddress + "/.well-known/acme-challenge/")
+	resp, err := http.Get("http://" + i.address + "/.well-known/acme-challenge/")
 	if err != nil {
 		err = errors.Wrap(err, "check for AES")
 		return
@@ -261,8 +330,8 @@ func (i *Installer) CheckAESServesACME(ipAddress string) (err error) {
 }
 
 // CheckAESHealth retrieves AES's idea of whether it is healthy, i.e. ready.
-func (i *Installer) CheckAESHealth(hostname string) error {
-	resp, err := http.Get("https://" + hostname + "/ambassador/v0/check_ready")
+func (i *Installer) CheckAESHealth() error {
+	resp, err := http.Get("https://" + i.hostname + "/ambassador/v0/check_ready")
 	if err != nil {
 		return err
 	}
@@ -279,8 +348,8 @@ func (i *Installer) CheckAESHealth(hostname string) error {
 // CheckHostnameFound tries to connect to check-blah.hostname to see whether DNS
 // has propagated. Each connect talks to a different hostname to try to avoid
 // NXDOMAIN caching.
-func (i *Installer) CheckHostnameFound(hostname string) error {
-	conn, err := net.Dial("tcp", fmt.Sprintf("check-%d.%s:443", time.Now().Unix(), hostname))
+func (i *Installer) CheckHostnameFound() error {
+	conn, err := net.Dial("tcp", fmt.Sprintf("check-%d.%s:443", time.Now().Unix(), i.hostname))
 	if err == nil {
 		conn.Close()
 	}
@@ -288,13 +357,13 @@ func (i *Installer) CheckHostnameFound(hostname string) error {
 }
 
 // CheckACMEIsDone queries the Host object and succeeds if its state is Ready.
-func (i *Installer) CheckACMEIsDone(hostname string) error {
-	state, err := i.CaptureKubectl("get Host state", "", "get", "host", hostname, "-o", "go-template={{.status.state}}")
+func (i *Installer) CheckACMEIsDone() error {
+	state, err := i.CaptureKubectl("get Host state", "", "get", "host", i.hostname, "-o", "go-template={{.status.state}}")
 	if err != nil {
 		return LoopFailedError(err.Error())
 	}
 	if state == "Error" {
-		reason, err := i.CaptureKubectl("get Host error", "", "get", "host", hostname, "-o", "go-template={{.status.errorReason}}")
+		reason, err := i.CaptureKubectl("get Host error", "", "get", "host", i.hostname, "-o", "go-template={{.status.errorReason}}")
 		if err != nil {
 			return LoopFailedError(err.Error())
 		}
@@ -309,9 +378,13 @@ func (i *Installer) CheckACMEIsDone(hostname string) error {
 			return errors.New("Waiting for NXDOMAIN retry")
 		}
 
-		i.show.Println("Acquiring TLS certificate via ACME has failed:")
-		i.show.Println(reason)
-		return LoopFailedError(fmt.Sprintf("ACME failed. More information: kubectl get host %s -o yaml", hostname))
+		// TODO: Windows incompatible, will not be bold but otherwise functions.
+		// TODO: rewrite Installer.show to make explicit calls to color.Bold.Printf(...) instead,
+		// TODO: along with logging.  Search for color.Bold to find usages.
+
+		i.show.Println()
+		i.show.Println(color.Bold.Sprintf("Acquiring TLS certificate via ACME has failed: %s", reason))
+		return LoopFailedError(fmt.Sprintf("ACME failed. More information: kubectl get host %s -o yaml", i.hostname))
 	}
 	if state != "Ready" {
 		return errors.Errorf("Host state is %s, not Ready", state)
@@ -364,23 +437,95 @@ func (i *Installer) GetInstalledImageVersion() (string, error) {
 }
 
 // Perform is the main function for the installer
-func (i *Installer) Perform(kcontext string) error {
+func (i *Installer) Perform(kcontext string) Result {
 	// Start
 	i.Report("install")
 
+	i.show.Println()
+	i.show.Println(color.Bold.Sprintf(welcomeInstall))
+
+	// Attempt to grab a reasonable default for the user's email address
+	defaultEmail, err := i.Capture("get email", true, "", "git", "config", "--global", "user.email")
+	if err != nil {
+		i.log.Print(err)
+		defaultEmail = ""
+	} else {
+		defaultEmail = strings.TrimSpace(defaultEmail)
+		if !validEmailAddress.MatchString(defaultEmail) {
+			defaultEmail = ""
+		}
+	}
+
+	// Ask for the user's email address
+	i.show.Println()
+	i.ShowWrapped(emailAsk)
+	// Do the goroutine dance to let the user hit Ctrl-C at the email prompt
+	gotEmail := make(chan (string))
+	var emailAddress string
+	go func() {
+		gotEmail <- getEmailAddress(defaultEmail, i.log)
+		close(gotEmail)
+	}()
+	select {
+	case emailAddress = <-gotEmail:
+		// Continue
+	case <-i.ctx.Done():
+		fmt.Println()
+		return UnhandledErrResult(errors.New("Interrupted"))
+	}
+	i.show.Println()
+
+	i.log.Printf("Using email address %q", emailAddress)
+
+	i.show.Println("========================================================================")
+	i.show.Println(beginningAESInstallation)
+	i.show.Println()
+
 	// Attempt to use kubectl
-	_, err := i.GetKubectlPath()
+	_, err = i.GetKubectlPath()
+	// err = errors.New("early error for testing")  // TODO: remove for production
 	if err != nil {
 		i.Report("fail_no_kubectl")
-		return fmt.Errorf(noKubectl)
+		err = browser.OpenURL(noKubectlURL)
+		return UnhandledErrResult(fmt.Errorf(noKubectl))
 	}
 
 	// Attempt to talk to the specified cluster
 	i.kubeinfo = k8s.NewKubeInfo("", kcontext, "")
 	if err := i.ShowKubectl("cluster-info", "", "cluster-info"); err != nil {
 		i.Report("fail_no_cluster")
-		return fmt.Errorf(noCluster)
+		err = browser.OpenURL(noClusterURL)
+		return UnhandledErrResult(fmt.Errorf(noCluster))
 	}
+	i.restConfig, err = i.kubeinfo.GetRestConfig()
+	if err != nil {
+		i.Report("fail_no_cluster")
+		return UnhandledErrResult(err)
+	}
+	i.coreClient, err = k8sClientCoreV1.NewForConfig(i.restConfig)
+	if err != nil {
+		i.Report("fail_no_cluster")
+		return UnhandledErrResult(err)
+	}
+
+	versions, err := i.CaptureKubectl("get versions", "", "version", "-o", "json")
+	if err != nil {
+		i.Report("fail_no_cluster")
+		return UnhandledErrResult(err)
+	}
+	kubernetesVersion := &kubernetesVersion{}
+	err = json.Unmarshal([]byte(versions), kubernetesVersion)
+	if err != nil {
+		// We tried to extract Kubernetes client and server versions but failed.
+		// This should not happen since we already validated the cluster-info, but still...
+		// It's not critical if this information is missing, other than for debugging purposes.
+		i.log.Printf("failed to read Kubernetes client and server versions: %v", err.Error())
+	}
+	i.k8sVersion = kubernetesVersion
+	// Metriton tries to parse fields with `version` in their keys and discards them if it can't.
+	// Using _v to keep the version value as string since Kubernetes versions vary in formats.
+	i.SetMetadatum("kubectl Version", "kubectl_v", i.k8sVersion.Client.GitVersion)
+	i.SetMetadatum("K8s Version", "k8s_v", i.k8sVersion.Server.GitVersion)
 
 	// Allow overriding the source domain (e.g., for smoke tests before release)
 	manifestsDomain := "www.getambassador.io"
@@ -395,12 +540,12 @@ func (i *Installer) Perform(kcontext string) error {
 	crdManifests, err := getManifest(fmt.Sprintf("https://%s/yaml/aes-crds.yaml", manifestsDomain))
 	if err != nil {
 		i.Report("fail_no_internet", ScoutMeta{"err", err.Error()})
-		return errors.Wrap(err, "download AES CRD manifests")
+		return UnhandledErrResult(errors.Wrap(err, "download AES CRD manifests"))
 	}
 	aesManifests, err := getManifest(fmt.Sprintf("https://%s/yaml/aes.yaml", manifestsDomain))
 	if err != nil {
 		i.Report("fail_no_internet", ScoutMeta{"err", err.Error()})
-		return errors.Wrap(err, "download AES manifests")
+		return UnhandledErrResult(errors.Wrap(err, "download AES manifests"))
 	}
 
 	// Figure out what version of AES is being installed
@@ -409,13 +554,13 @@ func (i *Installer) Perform(kcontext string) error {
 	if len(matches) != 2 {
 		i.log.Printf("matches is %+v", matches)
 		i.Report("fail_bad_manifests")
-		return errors.Errorf("Failed to parse downloaded manifests. Is there a proxy server interfering with HTTP downloads?")
+		return UnhandledErrResult(errors.Errorf("Failed to parse downloaded manifests. Is there a proxy server interfering with HTTP downloads?"))
 	}
-	aesVersion := matches[1]
-	i.SetMetadatum("AES version being installed", "aes_version", aesVersion)
+	i.version = matches[1]
+	i.SetMetadatum("AES version being installed", "aes_version", i.version)
 
 	// Display version information
-	i.ShowWrapped(fmt.Sprintf("-> Installing the Ambassador Edge Stack %s.", aesVersion))
+	i.ShowWrapped(fmt.Sprintf("-> Installing the Ambassador Edge Stack %s.", i.version))
 
 	// Try to determine cluster type from node labels
 	isKnownLocalCluster := false
@@ -466,56 +611,58 @@ func (i *Installer) Perform(kcontext string) error {
 			// Things will likely fail when we try to apply manifests
 		}
 		switch {
-		case aesVersion == installedVersion:
+		case i.version == installedVersion:
 			alreadyApplied = true
-			i.ShowWrapped(fmt.Sprintf("-> Found an existing installation of Ambassador Edge Stack %s.", aesVersion))
+			i.ShowWrapped(fmt.Sprintf("-> Found an existing installation of Ambassador Edge Stack %s.", i.version))
 		case installedVersion != "":
 			i.ShowWrapped(fmt.Sprintf("-> Found an existing installation of Ambassador Edge Stack %s.", installedVersion))
 			i.show.Println()
 			i.ShowWrapped(abortExisting)
 			i.show.Println()
 			i.ShowWrapped(seeDocs)
-			i.Report("fail_existing_aes", ScoutMeta{"installing", aesVersion}, ScoutMeta{"found", installedVersion})
-			return errors.Errorf("existing AES %s found when installing AES %s", installedVersion, aesVersion)
+			i.Report("fail_existing_aes", ScoutMeta{"installing", i.version}, ScoutMeta{"found", installedVersion})
+			return UnhandledErrResult(errors.Errorf("existing AES %s found when installing AES %s", installedVersion, i.version))
 		default:
-			i.ShowWrapped(abortCRDs)
-			i.show.Println()
-			i.ShowWrapped(seeDocs)
-			i.Report("fail_existing_crds")
-			return errors.New("CRDs found")
+			return Result{
+				Report:  "fail_existing_crds",
+				Message: abortCRDs,
+				URL:     seeDocsURL,
+				Err:     errors.New("CRDs found"),
+			}
 		}
 	}
 
 	if !alreadyApplied {
 		// Install the AES manifests
 
-		i.ShowWrapped("Downloading images. (This may take a minute.)")
+		i.ShowWrapped("-> Downloading images. (This may take a minute.)")
 
 		if err := i.ShowKubectl("install CRDs", crdManifests, "apply", "-f", "-"); err != nil {
 			i.Report("fail_install_crds")
-			return err
+			return UnhandledErrResult(err)
 		}
 
 		if err := i.ShowKubectl("wait for CRDs", "", "wait", "--for", "condition=established", "--timeout=90s", "crd", "-lproduct=aes"); err != nil {
 			i.Report("fail_wait_crds")
-			return err
+			return UnhandledErrResult(err)
 		}
 
 		if err := i.ShowKubectl("install AES", aesManifests, "apply", "-f", "-"); err != nil {
 			i.Report("fail_install_aes")
-			return err
+			return UnhandledErrResult(err)
 		}
 
 		if err := i.ShowKubectl("wait for AES", "", "-n", "ambassador", "wait", "--for", "condition=available", "--timeout=90s", "deploy", "-lproduct=aes"); err != nil {
 			i.Report("fail_wait_aes")
-			return err
+			return UnhandledErrResult(err)
 		}
 	}
 
 	// Wait for Ambassador Pod; grab AES install ID
+	i.show.Println("-> Checking the AES pod deployment")
 	if err := i.loopUntil("AES pod startup", i.GrabAESInstallID, lc2); err != nil {
 		i.Report("fail_pod_timeout")
-		return err
+		return UnhandledErrResult(err)
 	}
 
 	// Grab Helm information if present
@@ -530,185 +677,195 @@ func (i *Installer) Perform(kcontext string) error {
 	// Don't proceed any further if we know we are using a local (not publicly
 	// accessible) cluster. There's no point wasting the user's time on
 	// timeouts.
+
 	if isKnownLocalCluster {
 		i.Report("cluster_not_accessible")
 		i.show.Println("-> Local cluster detected. Not configuring automatic TLS.")
 		i.show.Println()
-		i.ShowWrapped(noTlsSuccess)
+		i.ShowWrapped(color.Bold.Sprintf(noTlsSuccess))
 		i.show.Println()
-		loginMsg := "Determine the IP address and port number of your Ambassador service.\n"
-		loginMsg += "(e.g., minikube service -n ambassador ambassador)\n"
-		loginMsg += fmt.Sprintf(loginViaIP, "IP_ADDRESS:PORT")
+		loginMsg := "Determine the IP address and port number of your Ambassador service, e.g.\n"
+		loginMsg += color.Bold.Sprintf("$ minikube service -n ambassador ambassador\n\n")
+		loginMsg += fmt.Sprintf(loginViaIP)
+		loginMsg += color.Bold.Sprintf("$ edgectl login -n ambassador IP_ADDRESS:PORT")
 		i.ShowWrapped(loginMsg)
 		i.show.Println()
 		i.ShowWrapped(seeDocs)
-		return nil
+		return UnhandledErrResult(nil)
 	}
 
-	// Grab load balancer IP address
-	i.ShowWrapped("-> Provisioning a cloud load balancer. (This may take a minute, depending on your cloud provider.)")
-	ipAddress := ""
-	grabIP := func() error {
-		var err error
-		ipAddress, err = i.GrabLoadBalancerIP()
-		return err
-	}
-	if err := i.loopUntil("Load Balancer", grabIP, lc5); err != nil {
+	// Grab load balancer address
+	i.show.Println("-> Provisioning a cloud load balancer")
+	if err := i.loopUntil("Load Balancer", i.GrabLoadBalancerAddress, lc5); err != nil {
 		i.Report("fail_loadbalancer_timeout")
 		i.show.Println()
 		i.ShowWrapped(failLoadBalancer)
 		i.show.Println()
-		i.ShowWrapped(noTlsSuccess)
+		i.ShowWrapped(color.Bold.Sprintf(noTlsSuccess))
 		i.ShowWrapped(seeDocs)
-		return err
+		return UnhandledErrResult(err)
 	}
 	i.Report("cluster_accessible")
-	i.show.Println("Your AES installation's IP address is", ipAddress)
+	i.show.Println("-> Your AES installation's address is", color.Bold.Sprintf(i.address))
 
 	// Wait for Ambassador to be ready to serve ACME requests.
-	if err := i.loopUntil("AES to serve ACME", func() error { return i.CheckAESServesACME(ipAddress) }, lc2); err != nil {
+	i.show.Println("-> Checking that AES is responding to ACME challenge")
+	if err := i.loopUntil("AES to serve ACME", i.CheckAESServesACME, lc2); err != nil {
 		i.Report("aes_listening_timeout")
-		i.ShowWrapped("It seems AES did not start in the expected time, or your IP address is not reachable from here.")
+		i.ShowWrapped("It seems AES did not start in the expected time, or the AES load balancer is not reachable from here.")
 		i.ShowWrapped(tryAgain)
-		i.ShowWrapped(noTlsSuccess)
+		i.ShowWrapped(color.Bold.Sprintf(noTlsSuccess))
 		i.ShowWrapped(seeDocs)
-		return err
+		return UnhandledErrResult(err)
 	}
 	i.Report("aes_listening")
 
 	i.show.Println("-> Automatically configuring TLS")
 
-	// Attempt to grab a reasonable default for the user's email address
-	defaultEmail, err := i.Capture("get email", "", "git", "config", "--global", "user.email")
-	if err != nil {
-		i.log.Print(err)
-		defaultEmail = ""
-	} else {
-		defaultEmail = strings.TrimSpace(defaultEmail)
-		if !validEmailAddress.MatchString(defaultEmail) {
-			defaultEmail = ""
-		}
-	}
-
-	// Ask for the user's email address
-	i.show.Println()
-	i.ShowWrapped(emailAsk)
-	// Do the goroutine dance to let the user hit Ctrl-C at the email prompt
-	gotEmail := make(chan (string))
-	var emailAddress string
-	go func() {
-		gotEmail <- getEmailAddress(defaultEmail, i.log)
-		close(gotEmail)
-	}()
-	select {
-	case emailAddress = <-gotEmail:
-		// Continue
-	case <-i.ctx.Done():
-		fmt.Println()
-		return errors.New("Interrupted")
-	}
-	i.show.Println()
-
-	i.log.Printf("Using email address %q", emailAddress)
-
-	// Send a request to acquire a DNS name for this cluster's IP address
+	// Send a request to acquire a DNS name for this cluster's load balancer
 	regURL := "https://metriton.datawire.io/register-domain"
+	regData := &registration{Email: emailAddress}
+	if !i.scout.Disabled() {
+		regData.AESInstallId = i.clusterID
+		regData.EdgectlInstallId = i.scout.installID
+	}
+	if net.ParseIP(i.address) != nil {
+		regData.Ip = i.address
+	} else {
+		regData.Hostname = i.address
+	}
 	buf := new(bytes.Buffer)
-	_ = json.NewEncoder(buf).Encode(registration{emailAddress, ipAddress})
+	_ = json.NewEncoder(buf).Encode(regData)
 	resp, err := http.Post(regURL, "application/json", buf)
 	if err != nil {
 		i.Report("dns_name_failure", ScoutMeta{"err", err.Error()})
-		return errors.Wrap(err, "acquire DNS name (post)")
+		return UnhandledErrResult(errors.Wrap(err, "acquire DNS name (post)"))
 	}
 	content, err := ioutil.ReadAll(resp.Body)
 	resp.Body.Close()
 	if err != nil {
 		i.Report("dns_name_failure", ScoutMeta{"err", err.Error()})
-		return errors.Wrap(err, "acquire DNS name (read body)")
+		return UnhandledErrResult(errors.Wrap(err, "acquire DNS name (read body)"))
 	}
 
 	if resp.StatusCode != 200 {
 		message := strings.TrimSpace(string(content))
 		i.Report("dns_name_failure", ScoutMeta{"code", resp.StatusCode}, ScoutMeta{"err", message})
 		i.show.Println("-> Failed to create a DNS name:", message)
-		i.show.Println()
-		i.ShowWrapped(noTlsSuccess)
-		i.ShowWrapped("If this IP address is reachable from here, you can access your installation without a DNS name.")
-		i.ShowWrapped(fmt.Sprintf(loginViaIP, ipAddress))
-		i.ShowWrapped(loginViaPortForward)
-		i.ShowWrapped(seeDocs)
-		return nil
+
+		userMessage := `
+<bold>Congratulations! You've successfully installed the Ambassador Edge Stack in your Kubernetes cluster. However, we cannot connect to your cluster from the Internet, so we could not configure TLS automatically.</>
+
+If this IP address is reachable from here, you can access your installation without a DNS name. The following command will open the Edge Policy Console once you accept a self-signed certificate in your browser.
+<bold>$ edgectl login -n ambassador {{ .address }}</>
+
+You can use port forwarding to access your Edge Stack installation and the Edge Policy Console.  You will need to accept a self-signed certificate in your browser.
+<bold>$ kubectl -n ambassador port-forward deploy/ambassador 8443 &</>
+<bold>$ edgectl login -n ambassador 127.0.0.1:8443</>
+`
+		return Result{
+			Message: userMessage,
+			URL:     seeDocsURL,
+			Report:  "", // FIXME: reported above due to additional metadata required
+		}
 	}
 
-	hostname := string(content)
-	i.show.Println("-> Acquiring DNS name", hostname)
+	i.hostname = string(content)
+	i.show.Println("-> Acquiring DNS name", color.Bold.Sprintf(i.hostname))
 
 	// Wait for DNS to propagate. This tries to avoid waiting for a ten
 	// minute error backoff if the ACME registration races ahead of the DNS
 	// name appearing for LetsEncrypt.
-	if err := i.loopUntil("DNS propagation to this host", func() error { return i.CheckHostnameFound(hostname) }, lc2); err != nil {
+	if err := i.loopUntil("DNS propagation to this host", i.CheckHostnameFound, lc2); err != nil {
 		i.Report("dns_name_propagation_timeout")
 		i.ShowWrapped("We are unable to resolve your new DNS name on this machine.")
 		i.ShowWrapped(seeDocs)
 		i.ShowWrapped(tryAgain)
-		return err
+		return UnhandledErrResult(err)
 	}
 	i.Report("dns_name_propagated")
 
 	// Create a Host resource
-	hostResource := fmt.Sprintf(hostManifest, hostname, hostname, emailAddress)
+	hostResource := fmt.Sprintf(hostManifest, i.hostname, i.hostname, emailAddress)
 	if err := i.ShowKubectl("install Host resource", hostResource, "apply", "-f", "-"); err != nil {
 		i.Report("fail_host_resource", ScoutMeta{"err", err.Error()})
 		i.ShowWrapped("We failed to create a Host resource in your cluster. This is unexpected.")
 		i.ShowWrapped(seeDocs)
-		return err
+		return UnhandledErrResult(err)
 	}
 
 	i.show.Println("-> Obtaining a TLS certificate from Let's Encrypt")
-	if err := i.loopUntil("TLS certificate acquisition", func() error { return i.CheckACMEIsDone(hostname) }, lc5); err != nil {
+	if err := i.loopUntil("TLS certificate acquisition", i.CheckACMEIsDone, lc5); err != nil {
 		i.Report("cert_provision_failed")
 		// Some info is reported by the check function.
 		i.ShowWrapped(seeDocs)
 		i.ShowWrapped(tryAgain)
-		return err
+		return UnhandledErrResult(err)
 	}
 	i.Report("cert_provisioned")
 	i.show.Println("-> TLS configured successfully")
-	if err := i.ShowKubectl("show Host", "", "get", "host", hostname); err != nil {
+	if err := i.ShowKubectl("show Host", "", "get", "host", i.hostname); err != nil {
 		i.ShowWrapped("We failed to retrieve the Host resource from your cluster that we just created. This is unexpected.")
 		i.ShowWrapped(tryAgain)
-		return err
+		return UnhandledErrResult(err)
 	}
 
 	i.show.Println()
-	i.ShowWrapped(fmt.Sprintf(fullSuccess, hostname))
+	i.show.Println("AES Installation Complete!")
+	i.show.Println("========================================================================")
+	i.show.Println()
+
+	// Show congratulations message
+	i.ShowWrapped(color.Bold.Sprintf(fullSuccess, i.hostname))
 	i.show.Println()
 
 	// Open a browser window to the Edge Policy Console
-	if err := do_login(i.kubeinfo, kcontext, "ambassador", hostname, false, false); err != nil {
-		return err
+	if err := do_login(i.kubeinfo, kcontext, "ambassador", i.hostname, true, true, false); err != nil {
+		return UnhandledErrResult(err)
 	}
 
-	if err := i.CheckAESHealth(hostname); err != nil {
+	// Show how to use edgectl login in the future
+	i.show.Println()
+	i.ShowWrapped(fmt.Sprintf(futureLogin, color.Bold.Sprintf("edgectl login "+i.hostname)))
+
+	if err := i.CheckAESHealth(); err != nil {
 		i.Report("aes_health_bad", ScoutMeta{"err", err.Error()})
 	} else {
 		i.Report("aes_health_good")
 	}
 
-	return nil
+	return UnhandledErrResult(nil)
 }
 
 // Installer represents the state of the installation process
 type Installer struct {
-	kubeinfo *k8s.KubeInfo
-	scout    *Scout
-	ctx      context.Context
-	cancel   context.CancelFunc
-	show     *log.Logger
-	log      *log.Logger
-	cmdOut   *log.Logger
-	cmdErr   *log.Logger
-	logName  string
+	// Cluster
+
+	kubeinfo   *k8s.KubeInfo
+	restConfig *rest.Config
+	coreClient *k8sClientCoreV1.CoreV1Client
+	k8sVersion *kubernetesVersion
+
+	// Reporting
+
+	scout *Scout
+
+	// Logging and management
+
+	ctx     context.Context
+	cancel  context.CancelFunc
+	show    *log.Logger
+	log     *log.Logger
+	cmdOut  *log.Logger
+	cmdErr  *log.Logger
+	logName string
+
+	// Install results
+
+	version   string // which AES is being installed
+	address   string // load balancer address
+	hostname  string // of the Host resource
+	clusterID string // the Ambassador unique clusterID
 }
 
 // NewInstaller returns an Installer object after setting up logging.
@@ -795,7 +952,25 @@ func (i *Installer) CaptureKubectl(name, input string, args ...string) (res stri
 		return
 	}
 	kargs = append([]string{kubectl}, kargs...)
-	return i.Capture(name, input, kargs...)
+	return i.Capture(name, true, input, kargs...)
+}
+
+// SilentCaptureKubectl calls kubectl and returns its stdout
+// without dumping all the output to the logger.
+func (i *Installer) SilentCaptureKubectl(name, input string, args ...string) (res string, err error) {
+	res = ""
+	kargs, err := i.kubeinfo.GetKubectlArray(args...)
+	if err != nil {
+		err = errors.Wrapf(err, "cluster access for %s", name)
+		return
+	}
+	kubectl, err := i.GetKubectlPath()
+	if err != nil {
+		err = errors.Wrapf(err, "kubectl not found %s", name)
+		return
+	}
+	kargs = append([]string{kubectl}, kargs...)
+	return i.Capture(name, false, input, kargs...)
 }
 
 // GetKubectlPath returns the full path to the kubectl executable, or an error if not found
@@ -803,15 +978,18 @@ func (i *Installer) GetKubectlPath() (string, error) {
 	return exec.LookPath("kubectl")
 }
 
-// Capture calls a command and returns its stdout, dumping all output to the
-// logger.
-func (i *Installer) Capture(name, input string, args ...string) (res string, err error) {
+// Capture calls a command and returns its stdout
+func (i *Installer) Capture(name string, logToStdout bool, input string, args ...string) (res string, err error) {
 	res = ""
 	resAsBytes := &bytes.Buffer{}
 	i.log.Printf("$ %s", strings.Join(args, " "))
 	cmd := exec.Command(args[0], args[1:]...)
 	cmd.Stdin = strings.NewReader(input)
-	cmd.Stdout = io.MultiWriter(NewLoggingWriter(i.cmdOut), resAsBytes)
+	if logToStdout {
+		cmd.Stdout = io.MultiWriter(NewLoggingWriter(i.cmdOut), resAsBytes)
+	} else {
+		cmd.Stdout = resAsBytes
+	}
 	cmd.Stderr = NewLoggingWriter(i.cmdErr)
 	err = cmd.Run()
 	if err != nil {
@@ -859,10 +1037,18 @@ func doWordWrap(text string, prefix string, lineWidth int) []string {
 	return lines
 }
 
+type kubernetesVersion struct {
+	Client k8sVersion.Info `json:"clientVersion"`
+	Server k8sVersion.Info `json:"serverVersion"`
+}
+
 // registration is used to register edgestack.me domains
 type registration struct {
-	Email string
-	Ip    string
+	Email            string
+	Ip               string
+	Hostname         string
+	EdgectlInstallId string
+	AESInstallId     string
 }
 
 const hostManifest = `
@@ -876,22 +1062,15 @@ spec:
     email: %s
 `
 
-const emailAsk = `Please enter an email address. We'll use this email address to notify you prior to domain and certificate expiration. We also share this email address with Let's Encrypt to acquire your certificate for TLS.`
+const welcomeInstall = "Installing the Ambassador Edge Stack"
 
-const loginViaIP = `
-The following command will open the Edge Policy Console once you accept a self-signed certificate in your browser.
+const emailAsk = `Please enter an email address for us to notify you before your TLS certificate and domain name expire. In order to acquire the TLS certificate, we share this email with Let’s Encrypt.`
 
-$ edgectl login -n ambassador %s
-` // ipAddress
+const beginningAESInstallation = "Beginning Ambassador Edge Stack Installation"
 
-const loginViaPortForward = `
-You can use port forwarding to access your Edge Stack installation and the Edge Policy Console.
+const loginViaIP = "The following command will open the Edge Policy Console once you accept a self-signed certificate in your browser.\n"
 
-$ kubectl -n ambassador port-forward deploy/ambassador 8443 &
-$ edgectl login -n ambassador 127.0.0.1:8443
-
-You will need to accept a self-signed certificate in your browser.
-`
+const loginViaPortForward = "You can use port forwarding to access your Edge Stack installation and the Edge Policy Console.  You will need to accept a self-signed certificate in your browser.\n"
 
 const failLoadBalancer = `
 Timed out waiting for the load balancer's IP address for the AES Service.
@@ -904,7 +1083,7 @@ const tryAgain = "If this appears to be a transient failure, please try running 
 const abortExisting = `
 This tool does not support upgrades/downgrades at this time.
 
-Aborting the installer to avoid corrupting an existing installation of AES.
+The installer will now quit to avoid corrupting an existing installation of AES.
 `
 
 const abortCRDs = `
@@ -915,24 +1094,31 @@ Removing the CRDs will cause your existing Ambassador Mappings and other resourc
 
 $ kubectl delete crd -l product=aes
 
-Aborting the installer to avoid corrupting an existing (but undetected) installation.
+The installer will now quit to avoid corrupting an existing (but undetected) installation.
 `
+const seeDocsURL = "https://www.getambassador.io/docs/latest/tutorials/getting-started/"
+const seeDocs = "See " + seeDocsURL
 
-const seeDocs = "See https://www.getambassador.io/docs/latest/tutorials/getting-started/"
+const phoneHomeDisabled = "INFO: phone-home is disabled by environment variable"
+const installAndTraceIDs = "INFO: install_id = %s; trace_id = %s"
 
 var validEmailAddress = regexp.MustCompile("^[a-zA-Z0-9.!#$%&'*+/=?^_`{|}~-]+@[a-zA-Z0-9](?:[a-zA-Z0-9-]{0,61}[a-zA-Z0-9])?(?:\\.[a-zA-Z0-9](?:[a-zA-Z0-9-]{0,61}[a-zA-Z0-9])?)*$")
 
-const fullSuccess = "Congratulations! You've successfully installed the Ambassador Edge Stack in your Kubernetes cluster. Visit %s to access your Edge Stack installation and for additional configuration." // hostname
+const fullSuccess = `Congratulations! You've successfully installed the Ambassador Edge Stack in your Kubernetes cluster. Visit https://%s` // hostname
+
+const futureLogin = `In the future, to log in to the Ambassador Edge Policy Console, run
+$ %s` // "edgectl login <hostname>"
 
 const noTlsSuccess = "Congratulations! You've successfully installed the Ambassador Edge Stack in your Kubernetes cluster. However, we cannot connect to your cluster from the Internet, so we could not configure TLS automatically."
 
+const noKubectlURL = "https://kubernetes.io/docs/tasks/tools/install-kubectl/"
 const noKubectl = `
 The installer depends on the 'kubectl' executable. Make sure you have the latest release downloaded in your PATH, and that you have executable permissions.
+Visit ` + noKubectlURL + ` for more information and instructions.`
 
-Visit https://kubernetes.io/docs/tasks/tools/install-kubectl/ for more information and instructions.`
-
+const noClusterURL = "https://kubernetes.io/docs/setup/"
 const noCluster = `
 Unable to communicate with the remote Kubernetes cluster using your kubectl context.
 
 To further debug and diagnose cluster problems, use 'kubectl cluster-info dump' 
-or get started and run Kubernetes https://kubernetes.io/docs/setup/`
+or get started and run Kubernetes: ` + noClusterURL